Re implement Asset_DAO; Refactore app rotes
